requirements.txt
bottle==0.12.10
line_bot_sdk==1.17.0
google_api_python_client==1.7.9
twilio==6.5.1
pusher==1.7.2
GoogleAppEngineCloudStorageClient==1.9.22.1
oauth2client==4.1.3
Arpeggio=1.9.2
